Fix InputArugment implicit operator throwing exception in ScriptSoundId

+++ source/scripting_v3/GTA/ScriptSoundId.cs
@@ -174,7 +174,7 @@ namespace GTA
 
 		public static implicit operator InputArgument(ScriptSoundId value)
 		{
-			return new InputArgument(value.Id);
+			return new InputArgument((ulong)value.Id);
 		}
 		public static implicit operator ScriptSoundId(int value)
 		{
